requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
numpy
requests
pydantic
